+++ CHANGELOG.md
@@ -8,4 +8,9 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- `cli.js list-meeting` command to list the meetings running on a BBB instance
+- `cli.js stress` command to start a stress test on a BBB server
+
 [Unreleased]: https://github.com/openfun/bbb-stress-test